backend/pyproject.toml
[project]
name = "content-embedding-storage"
version = "0.1.0"
description = "Content extraction, embedding generation, and vector storage system"
authors = [
    {name = "Claude", email = "claude@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "cohere>=5.0.0",
    "qdrant-client>=1.8.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]